Ejercicio/website/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
sentence-transformers==2.5.1
faiss-cpu==1.7.4
openai==1.3.0
python-dotenv==1.0.0
tiktoken==0.5.2
numpy==1.26.2
